backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==11.0.0
tensorflow==2.18.0
numpy>=2.0.0,<3.0.0
scikit-learn>=1.4.2,<2.0.0
python-dotenv==1.0.0